extracted schema insides into the module, added environment to the results object
